requirements.txt
streamlit
requests
pandas
python-dotenv
google-generativeai 
supabase         
altair               
streamlit-autorefresh 
wheel